requirements.txt
pytest>=7.0.0
pyodbc-5.3.0
numpy-2.3.5
pandas-2.3.3
python-dateutil-2.9.0.post0
pytz-2025.2
six-1.17.0
tzdata-2025.2
google-cloud-storage==2.14.0